+++ skforecast/ForecasterAutoregCustom/tests/test_predict.py
@@ -1,3 +1,5 @@
+# Unit test predict ForecasterAutoregCustom
+# ==============================================================================
 from pytest import approx
 import numpy as np
 import pandas as pd