+++ skforecast/ForecasterAutoregDirect/tests/test_create_lags.py
@@ -1,9 +1,9 @@
-# Unit test _create_lags ForecasterAutoregMultiOutput
+# Unit test _create_lags ForecasterAutoregDirect
 # ==============================================================================
 import pytest
 import numpy as np
 import pandas as pd
-from skforecast.ForecasterAutoregMultiOutput import ForecasterAutoregMultiOutput
+from skforecast.ForecasterAutoregDirect import ForecasterAutoregDirect
 from sklearn.linear_model import LinearRegression
 
     
@@ -12,7 +12,7 @@ def test_create_lags_exception_when_len_of_y_is_lower_than_maximum_lag():
     Test exception is raised when length of y is lower than maximum lag included
     in the forecaster.
     '''
-    forecaster = ForecasterAutoregMultiOutput(LinearRegression(), lags=10, steps=2)
+    forecaster = ForecasterAutoregDirect(LinearRegression(), lags=10, steps=2)
     with pytest.raises(Exception):
         forecaster._create_lags(y=np.arange(5))
 
@@ -23,7 +23,7 @@ def test_create_lags_when_lags_is_3_steps_1_and_y_is_numpy_arange_10():
     Test matrix of lags created properly when langs is 3, steps is 1 and y is
     np.arange(10).
     '''
-    forecaster = ForecasterAutoregMultiOutput(LinearRegression(), lags=3, steps=1)
+    forecaster = ForecasterAutoregDirect(LinearRegression(), lags=3, steps=1)
     results = forecaster._create_lags(y=np.arange(10))
     expected = (np.array([[2., 1., 0.],
                           [3., 2., 1.],
@@ -49,7 +49,7 @@ def test_create_lags_when_lags_is_3_steps_2_and_y_is_numpy_arange_10():
     Test matrix of lags created properly when langs is 3, steps is 2 and y is
     np.arange(10).
     '''
-    forecaster = ForecasterAutoregMultiOutput(LinearRegression(), lags=3, steps=2)
+    forecaster = ForecasterAutoregDirect(LinearRegression(), lags=3, steps=2)
     results = forecaster._create_lags(y=np.arange(10))
     expected = (np.array([[2., 1., 0.],
                           [3., 2., 1.],
@@ -73,7 +73,7 @@ def test_create_lags_when_lags_is_3_steps_5_and_y_is_numpy_arange_10():
     Test matrix of lags created properly when langs is 3, steps is 5 and y is
     np.arange(10).
     '''
-    forecaster = ForecasterAutoregMultiOutput(LinearRegression(), lags=3, steps=5)
+    forecaster = ForecasterAutoregDirect(LinearRegression(), lags=3, steps=5)
     results = forecaster._create_lags(y=np.arange(10))
     expected = (np.array([[2., 1., 0.],
                           [3., 2., 1.],