+++ skforecast/ForecasterAutoregDirect/tests/test_predict.py
@@ -1,9 +1,8 @@
-# Unit test predict ForecasterAutoreg
+# Unit test predict ForecasterAutoregDirect
 # ==============================================================================
-from pytest import approx
 import numpy as np
 import pandas as pd
-from skforecast.ForecasterAutoreg import ForecasterAutoreg
+from skforecast.ForecasterAutoregDirect import ForecasterAutoregDirect
 from sklearn.linear_model import LinearRegression
 
         
@@ -11,12 +10,13 @@ def test_predict_output_when_regressor_is_LinearRegression():
     '''
     Test predict output when using LinearRegression as regressor.
     '''
-    forecaster = ForecasterAutoreg(LinearRegression(), lags=3)
+    forecaster = ForecasterAutoregDirect(LinearRegression(), lags=3, steps=3)
     forecaster.fit(y=pd.Series(np.arange(50)))
-    predictions = forecaster.predict(steps=5)
+    results = forecaster.predict()
     expected = pd.Series(
-                data = np.array([50., 51., 52., 53., 54.]),
-                index = pd.RangeIndex(start=50, stop=55, step=1),
+                data = np.array([50., 51., 52.]),
+                index = pd.RangeIndex(start=50, stop=53, step=1),
                 name = 'pred'
                )
-    pd.testing.assert_series_equal(predictions, expected)
+    pd.testing.assert_series_equal(results, expected)
+