+++ skforecast/ForecasterAutoregDirect/tests/test_set_params.py
@@ -1,12 +1,12 @@
-# Unit test set_params ForecasterAutoregMultiOutput
+# Unit test set_params ForecasterAutoregDirect
 # ==============================================================================
-from skforecast.ForecasterAutoregMultiOutput import ForecasterAutoregMultiOutput
+from skforecast.ForecasterAutoregDirect import ForecasterAutoregDirect
 from sklearn.linear_model import LinearRegression
 
 
 def test_set_params():
     
-    forecaster = ForecasterAutoregMultiOutput(LinearRegression(fit_intercept=True), lags=3, steps=3)
+    forecaster = ForecasterAutoregDirect(LinearRegression(fit_intercept=True), lags=3, steps=3)
     new_params = {'fit_intercept': False}
     forecaster.set_params(**new_params)
     expected = {'copy_X': True,