+++ skforecast/model_selection/tests/test_time_series_splitter.py
@@ -1,20 +1,23 @@
+# Unit test time_series_splitter
+# ==============================================================================
 import pytest
 import numpy as np
 from skforecast.model_selection import time_series_splitter
 
-# Test test_time_series_splitter
-#-------------------------------------------------------------------------------
+
 def test_time_series_splitter_exception_when_y_is_numpy_array_with_more_than_1_dimesion():
 
     results = time_series_splitter(np.arange(10).reshape(-1, 2), initial_train_size=3, steps=1)
     with pytest.raises(Exception):
        list(results)
        
+
 def test_time_series_splitter_exception_when_y_is_list():
 
     results = time_series_splitter([0,1,2,3,4], initial_train_size=3, steps=1)
     with pytest.raises(Exception):
        list(results)
+
 
 def test_time_series_splitter_when_y_is_numpy_arange_10_initial_train_size_5_steps_1_allow_incomplete_fold_True():
 
@@ -33,6 +36,7 @@ def test_time_series_splitter_when_y_is_numpy_arange_10_initial_train_size_5_ste
                 (range(0, 9), range(9, 10))]
     assert results == expected
 
+
 def test_time_series_splitter_when_y_is_numpy_arange_10_initial_train_size_5_steps_5_allow_incomplete_fold_True():
 
     results = time_series_splitter(
@@ -46,6 +50,7 @@ def test_time_series_splitter_when_y_is_numpy_arange_10_initial_train_size_5_ste
     expected = [(range(0, 5), range(5, 10))]
     assert results == expected
 
+
 def test_time_series_splitter_when_y_is_numpy_arange_10_initial_train_size_5_steps_3_allow_incomplete_fold_False():
 
     results = time_series_splitter(
@@ -58,6 +63,7 @@ def test_time_series_splitter_when_y_is_numpy_arange_10_initial_train_size_5_ste
     results = list(results)
     expected = [(range(0, 5), range(5, 8))]
     assert results == expected
+
 
 def test_time_series_splitter_when_y_is_numpy_arange_10_initial_train_size_5_steps_3_allow_incomplete_fold_True():
 
@@ -86,6 +92,7 @@ def test_time_series_splitter_when_y_is_numpy_arange_10_initial_train_size_5_ste
     results = list(results)
     expected = []
     assert results == expected
+
 
 def test_time_series_splitter_when_y_is_numpy_arange_10_initial_train_size_5_steps_20_allow_incomplete_fold_True():